[iOS] Combine BraveWebView & TabWebView
This combines the two web view types we have: `BraveWebView` & `TabWebView`, and removes the tab reference from the `TabWebView` as it wasn't needed
